requirements.txt
Flask==2.0.3
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
mysqlclient==2.0.3
python-dotenv==0.19.2
flask-bcrypt